requirements.txt
python==3.12.0
python-dotenv==1.1.0
pyyaml==6.0.2
openai==1.70.0